+++ CHANGES.md
@@ -6,6 +6,7 @@
 * re-use valid cache timeout
 * refactor getting cache key to `get_cache_key()`
 * ensure newest tag present in release assets array
+* check to see `$existing_cache` timeout is valid
 
 #### 12.23.1 / 2026-02-12
 * add guard to release asset development download link in REST API